Move fclose_exclusive and fopen_exclusive into ofstream_wrapper_exclusive
`fopen_exclusive` is now `ofstream_wrapper_exclusive::open` with the additional code for passing the mode and the error reporting taken from the constructor of `ofstream_wrapper_exclusive`.

`fopen_exclusive` is now `ofstream_wrapper_exclusive::open` with additional code from the original `close` function for error reporting and for moving the temporary file to its final destination.
