+++ src/mapsharing.h
@@ -44,9 +44,5 @@ void setDefaults();
 int getLock( const char *lockName );
 void releaseLock( int fd, const char *lockName );
 extern std::map<std::string, int> lockFiles;
-void fopen_exclusive( std::ofstream &fout, const char *filename,
-                      std::ios_base::openmode mode = std::ios_base::out );
-//std::ofstream fopen_exclusive(const char* filename);
-void fclose_exclusive( std::ofstream &fout, const char *filename );
 
 #endif
